compilation fixes for Python 2.5
git-svn-id: https://svn.grrrr.org/ext/py/trunk@2384 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/pybase.cpp
@@ -678,7 +678,7 @@ void pybase::Reload()
 
     if(reg) {
         PyObject *key;
-        int pos = 0;
+        Py_ssize_t pos = 0;
         while(PyDict_Next(reg,&pos,&key,NULL)) {
             pybase *th = (pybase *)PyLong_AsLong(key);
             FLEXT_ASSERT(th);
@@ -697,7 +697,7 @@ void pybase::Reload()
             SetRegistry(REGNAME,reg);
 
             PyObject *key;
-            int pos = 0;
+            Py_ssize_t pos = 0;
             while(PyDict_Next(reg,&pos,&key,NULL)) {
                 pybase *th = (pybase *)PyLong_AsLong(key);
                 FLEXT_ASSERT(th);
